utility: specialise get_dataspace etc. for h5xx::attribute
H5Aget_space must be used only for attributes. This is a preliminary
solution only. A better solution will be a class h5xx::dataspace which
constructible from either an attribute or a dataset.
